refactor: move epicycle core to separate package
