+++ src/window/__init__.py
@@ -1,0 +1,3 @@
+"""PyGame views and controllers."""
+
+from .main import start
