+++ src/window/commands/delete_wrapper.py
@@ -1,6 +1,5 @@
-from state import State
-
 from epicycle.epicycle_wrapper import EpicycleWrapper
+from state import State
 
 from .interface import ICommand
