+++ src/window/commands/pause.py
