+++ src/window/handlers/exit_handler.py
@@ -2,15 +2,14 @@ import pygame as pg
 from pygame.event import Event
 from state import State
 
-from epicycle.commands import ICommand, PauseCommand
+from window.commands import ExitCommand, ICommand
 
 from .interface import IHandler
 
 
-class PauseHandler(IHandler):
+class ExitHandler(IHandler):
     @classmethod
     def handle(cls, event: Event, state: State) -> ICommand | None:
-        if event.type == pg.KEYDOWN:
-            if event.key == pg.K_p:
-                return PauseCommand(state)
+        if event.type == pg.QUIT:
+            return ExitCommand()
         return None
