+++ src/window/handlers/interface.py
@@ -3,7 +3,7 @@ import abc
 import pygame as pg
 from state import State
 
-from epicycle.commands import ICommand
+from window.commands import ICommand
 
 
 class IHandler(abc.ABC):
